Add `score_string` & `decrypt_string` utils to their respective routes
